Moved files around to get autoloaded by rails and removed the proactiveautoloader
