+++ lib/report/chainable.rb
