+++ lib/report/controller.rb
