+++ lib/report/filter.rb
@@ -1,9 +1,6 @@
 require "set"
 
 class Report::Filter
-  extend ProactiveAutoloader
-  autoload :Base, 'reporting_engine/report/filter/base'
-  autoload :NoFilter, 'reporting_engine/report/filter/no_filter'
 
   def self.all
     @all ||= Set[]
