+++ lib/report/filter/multi_choice.rb
