+++ lib/report/filter/no_filter.rb
