+++ lib/report/group_by.rb
@@ -1,12 +1,7 @@
 require "set"
 
 class Report::GroupBy
-  extend ProactiveAutoloader
   include Report::QueryUtils
-  autoload :Base, 'reporting_engine/report/group_by/base'
-  autoload :RubyAggregation, 'reporting_engine/report/group_by/ruby_aggregation'
-  autoload :SingletonValue, 'reporting_engine/report/group_by/singleton_value.rb'
-  autoload :SqlAggregation, 'reporting_engine/report/group_by/sql_aggregation'
 
   def self.all
     Set[engine::GroupBy::SingletonValue]
