+++ lib/report/group_by/base.rb
