+++ lib/report/group_by/ruby_aggregation.rb
