+++ lib/report/group_by/singleton_value.rb
