+++ lib/report/group_by/sql_aggregation.rb
