+++ lib/report/operator.rb
