+++ lib/report/query_utils.rb
