+++ lib/report/result.rb
