+++ lib/report/table.rb
