+++ lib/report/transformer.rb
