+++ lib/report/validation/integers.rb
