+++ lib/report/validation/sql.rb
