+++ lib/report/walker.rb
