+++ lib/widget/base.rb
