+++ lib/widget/controls.rb
@@ -1,5 +1,4 @@
 class Widget::Controls < Widget::Base
-  extend ProactiveAutoloader
 
   def cache_key
     "#{super}#{@subject.new_record? ? 1 : 0}"
