+++ lib/widget/controls/apply.rb
