+++ lib/widget/controls/clear.rb
