+++ lib/widget/controls/delete.rb
