+++ lib/widget/controls/query_name.rb
