+++ lib/widget/filters/base.rb
