+++ lib/widget/filters/multi_choice.rb
@@ -1,4 +1,6 @@
-
+#make sure to require Widget::Filters::Base first because otherwise
+#ruby might find Base within Widget and Rails will not load it
+require_dependency 'widget/filters/base'
 class Widget::Filters::MultiChoice < Widget::Filters::Base
 
   def render
