+++ lib/widget/help.rb
