+++ lib/widget/settings.rb
