+++ lib/widget/settings/fieldset.rb
