+++ lib/widget/table.rb
@@ -1,5 +1,3 @@
-require_dependency 'reporting_engine/report/inherited_attribute'
-
 class Widget::Table < Widget::Base
   extend Report::InheritedAttribute
   include ReportingHelper
