+++ lib/widget/table/progressbar.rb
