+++ lib/widget/table/report_table.rb
